Move implementations to ConvDiff.cpp.
Add explicit instantiations of function templates for Schemes::CD and Schemes::SG.
